+++ GPU/Directx9/DepalettizeShaderDX9.cpp
@@ -24,7 +24,7 @@
 #include "GPU/Directx9/DepalettizeShaderDX9.h"
 #include "GPU/Common/DepalettizeShaderCommon.h"
 #include "gfx/d3d9_shader.h"
-#include "GPU/Directx9/helper/dx_state.h"
+#include "gfx/d3d9_state.h"
 
 namespace DX9 {
