+++ GPU/Directx9/StateMappingDX9.cpp
@@ -24,7 +24,7 @@
 #include "Core/Reporting.h"
 
 #include "gfx/d3d9_shader.h"
-#include "helper/dx_state.h"
+#include "gfx/d3d9_state.h"
 
 #include "GPU/Directx9/GPU_DX9.h"
 #include "GPU/Directx9/ShaderManagerDX9.h"
