+++ ext/native/native.vcxproj.filters
@@ -314,6 +314,9 @@
     <ClInclude Include="gfx\d3d9_shader.h">
       <Filter>gfx</Filter>
     </ClInclude>
+    <ClInclude Include="gfx\d3d9_state.h">
+      <Filter>gfx</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="gfx\gl_debug_log.cpp">
@@ -755,6 +758,9 @@
       <Filter>thin3d</Filter>
     </ClCompile>
     <ClCompile Include="gfx\d3d9_shader.cpp">
+      <Filter>gfx</Filter>
+    </ClCompile>
+    <ClCompile Include="gfx\d3d9_state.cpp">
       <Filter>gfx</Filter>
     </ClCompile>
   </ItemGroup>
